requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
discord.py
aiohttp
openai
